refactor: split test file
